Refactor hazards sidebar
